Extract enterprise audit event payload module
